Move PathType to Readers namespace
