+++ src/EventLogExpert.Eventing.Tests/EventResolvers/EventXmlResolverTests.cs
@@ -2,8 +2,8 @@
 // // Licensed under the MIT License.
 
 using EventLogExpert.Eventing.EventResolvers;
-using EventLogExpert.Eventing.Helpers;
 using EventLogExpert.Eventing.Models;
+using EventLogExpert.Eventing.Readers;
 
 namespace EventLogExpert.Eventing.Tests.EventResolvers;
