+++ src/EventLogExpert.Eventing/Models/EventRecord.cs
@@ -1,7 +1,7 @@
-﻿// // Copyright (c) Microsoft Corporation.
+// // Copyright (c) Microsoft Corporation.
 // // Licensed under the MIT License.
 
-using EventLogExpert.Eventing.Helpers;
+using EventLogExpert.Eventing.Readers;
 using System.Security.Principal;
 
 namespace EventLogExpert.Eventing.Models;
