+++ src/EventLogExpert.UI.Tests/TestUtils/EventUtils.cs
@@ -1,8 +1,8 @@
-﻿// // Copyright (c) Microsoft Corporation.
+// // Copyright (c) Microsoft Corporation.
 // // Licensed under the MIT License.
 
-using EventLogExpert.Eventing.Helpers;
 using EventLogExpert.Eventing.Models;
+using EventLogExpert.Eventing.Readers;
 
 namespace EventLogExpert.UI.Tests.TestUtils;
