+++ src/EventLogExpert/_Imports.razor
@@ -18,6 +18,7 @@
 @using EventLogExpert.Components.Modals.Filters
 @using EventLogExpert.Components.Sections
 @using EventLogExpert.Eventing.Helpers
+@using EventLogExpert.Eventing.Readers
 @using EventLogExpert.UI
 @using EventLogExpert.UI.Models
